requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-multiselectfield==0.1.13
django-rest-framework==0.1.0
djangorestframework==3.15.2
multiselect==0.1.0
pillow==10.4.0
sqlparse==0.5.1
tzdata==2024.1
